made changes based on comments, remade SwitchToMagazineTest
