+++ HomeWork/HomeWork/Tests/SearchFieldTest.cs
@@ -1,0 +1,33 @@
+﻿using NUnit.Framework;
+using OpenQA.Selenium;
+
+namespace HomeWork
+{
+    internal class SearchFieldTest
+    {
+        private IWebDriver driver;
+        private UserService service;
+
+        [SetUp]
+        public void Setup()
+        {
+            driver = new OpenQA.Selenium.Chrome.ChromeDriver();
+            driver.Navigate().GoToUrl("https://ek.ua/");
+            driver.Manage().Window.Maximize();
+            service = new UserService(driver);
+        }
+
+        [Test]
+        public void Test1()
+        {
+            service.Search("iPhone 13 Pro 256");
+        }
+
+        [TearDown]
+        public void Test2()
+        {
+            driver.Quit();
+            driver.Dispose();
+        }
+    }
+}
