+++ HomeWork/HomeWork/Tests/SwitchToMagazineTest.cs
