+++ HomeWork/HomeWork/bin/Debug/net6.0/nunit_random_seed.tmp
@@ -1,1 +1,1 @@
-1439127288
+243785127
